+++ plugins/postcss-oklab-function/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Changes to PostCSS OKLab Function
 
+### Unreleased (minor)
+
+- Allow percentage and number units in more color components.
+
+```css
+.percentages {
+	color-1: oklab(40% 0.309% 0.975%);
+	color-2: oklch(40% 31.718385875% 34.568626);
+
+	/* becomes */
+
+	color-1: rgb(73, 71, 69);
+	color-1: color(display-p3 0.28515 0.27983 0.27246);
+	color-2: rgb(126, 37, 15);
+	color-2: color(display-p3 0.45368 0.16978 0.09411);
+}
+
+.numbers {
+	color-1: oklab(0.40 0.001236 0.0039);
+	color-2: oklch(0.40 0.1268735435 34.568626);
+
+	/* becomes */
+
+	color-1: rgb(73, 71, 69);
+	color-1: color(display-p3 0.28515 0.27983 0.27246);
+	color-2: rgb(126, 37, 15);
+	color-2: color(display-p3 0.45368 0.16978 0.09411);
+}
+```
+
 ### 1.0.1 (February 12, 2022)
 
 - Updated `@csstools/postcss-progressive-custom-properties` to `1.1.0`.