+++ plugins/postcss-oklab-function/CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Unreleased (patch)
 
 - Fix gamut mapping giving overly unsaturated colors.
-- Implement powerless color components.
+- Implement powerless color components in gamut mapping.
 
 ### 1.0.1 (February 12, 2022)