+++ plugins/postcss-oklab-function/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to PostCSS OKLab Function
 
+### Unreleased (patch)
+
+- Fix gamut mapping giving overly unsaturated colors.
+- Implement powerless color components in gamut mapping.
+
 ### 1.0.1 (February 12, 2022)
 
 - Updated `@csstools/postcss-progressive-custom-properties` to `1.1.0`.