+++ plugins/postcss-oklab-function/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changes to PostCSS OKLab Function
 
-### Unreleased (patch)
+### 1.0.2 (March 8, 2022)
 
 - Fix gamut mapping giving overly unsaturated colors.
 - Implement powerless color components in gamut mapping.